+++ tests/test_modeling_common.py
@@ -34,9 +34,7 @@ from pytest import mark
 
 from transformers import (
     AutoModel,
-    AutoModelForCausalLM,
     AutoModelForSequenceClassification,
-    DataCollatorWithFlattening,
     PretrainedConfig,
     PreTrainedModel,
     is_torch_available,
@@ -125,7 +123,6 @@ if is_accelerate_available():
 
 if is_torch_available():
     import torch
-    import torch.nn.functional as F
     from safetensors.torch import load_file as safe_load_file
     from safetensors.torch import save_file as safe_save_file
     from torch import nn
@@ -233,6 +230,11 @@ def _test_eager_matches_sdpa_inference(
     for model_class in self.all_model_classes:
         config, inputs_dict = self.model_tester.prepare_config_and_inputs_for_common()
         set_config_for_less_flaky_test(config)
+
+        # If it's a model with sliding window attention, let's test it with sliding window
+        if hasattr(config, "sliding_window"):
+            config.sliding_window = 2
+
         model = model_class(config)
         # TODO: standardize the interfaces for musicgen models, see other todo in this test
         if model.__class__.__name__ == "MusicgenMelodyForConditionalGeneration":
@@ -1182,41 +1184,6 @@ class ModelTesterMixin:
             inputs = self._prepare_for_class(inputs_dict, model_class, return_labels=True)
             loss = model(**inputs).loss
             loss.backward()
-
-    def test_causal_lm_can_accept_kwargs(self):
-        if not getattr(self.model_tester, "is_training", False):
-            self.skipTest(reason="ModelTester is not configured to run training tests")
-
-        valid_model_class = False
-        incompatible_models = (
-            "MusicgenForCausalLM",
-            "MusicgenMelodyForCausalLM",
-            "MllamaForCausalLM",
-            "CpmAntForCausalLM",
-            "GotOcr2ForConditionalGeneration",
-        )
-        for model_class in self.all_model_classes:
-            if (
-                model_class.__name__ in get_values(MODEL_FOR_CAUSAL_LM_MAPPING_NAMES)
-                and model_class.__name__ not in incompatible_models
-            ):
-                valid_model_class = True
-        if not valid_model_class:
-            self.skipTest(reason="No causal lm model classes found")
-        for model_class in self.all_model_classes:
-            model_name = model_class.__name__
-            if model_name in get_values(MODEL_FOR_CAUSAL_LM_MAPPING_NAMES) and model_name not in incompatible_models:
-                config, inputs_dict = self.model_tester.prepare_config_and_inputs_for_common()
-
-                with tempfile.TemporaryDirectory() as tmpdir:
-                    with torch.device(torch_device):
-                        model_eager = AutoModelForCausalLM.from_config(config, dtype=torch.float32)
-
-                    model_eager.save_pretrained(tmpdir)
-                    model = AutoModelForCausalLM.from_pretrained(tmpdir, dtype=torch.float32, device_map=torch_device)
-                    inputs_dict["num_items_in_batch"] = torch.tensor(inputs_dict["input_ids"].shape[0])
-                    inputs_dict["labels"] = inputs_dict["input_ids"]
-                    _ = model(**inputs_dict, return_dict=False)
 
     def test_training_gradient_checkpointing(self):
         # Scenario - 1 default behaviour
@@ -2744,65 +2711,6 @@ class ModelTesterMixin:
                     model, tuple_inputs, dict_inputs, {"output_hidden_states": True, "output_attentions": True}
                 )
 
-    # Don't copy this method to model specific test file!
-    # TODO: remove this method once the issues are all fixed!
-    def _make_attention_mask_non_null(self, inputs_dict):
-        """Make sure no sequence has all zeros as attention mask"""
-
-        for k in ["attention_mask", "encoder_attention_mask", "decoder_attention_mask"]:
-            if k in inputs_dict:
-                attention_mask = inputs_dict[k]
-
-                # Make sure no all 0s attention masks - to avoid failure at this moment.
-                # Put `1` at the beginning of sequences to make it still work when combining causal attention masks.
-                # TODO: remove this line once a fix regarding large negative values for attention mask is done.
-                attention_mask = torch.cat(
-                    [torch.ones_like(attention_mask[:, :1], dtype=attention_mask.dtype), attention_mask[:, 1:]], dim=-1
-                )
-
-                # Here we make the first sequence with all 0s as attention mask.
-                # Currently, this will fail for `TFWav2Vec2Model`. This is caused by the different large negative
-                # values, like `1e-4`, `1e-9`, `1e-30` and `-inf` for attention mask across models/frameworks.
-                # TODO: enable this block once the large negative values thing is cleaned up.
-                # (see https://github.com/huggingface/transformers/issues/14859)
-                # attention_mask = torch.cat(
-                #     [torch.zeros_like(attention_mask[:1], dtype=attention_mask.dtype), attention_mask[1:]],
-                #     dim=0
-                # )
-
-                inputs_dict[k] = attention_mask
-
-    # Don't copy this method to model specific test file!
-    # TODO: remove this method once the issues are all fixed!
-    def _postprocessing_to_ignore_test_cases(self, tf_outputs, pt_outputs, model_class):
-        """For temporarily ignoring some failed test cases (issues to be fixed)"""
-
-        tf_keys = {k for k, v in tf_outputs.items() if v is not None}
-        pt_keys = {k for k, v in pt_outputs.items() if v is not None}
-
-        key_differences = tf_keys.symmetric_difference(pt_keys)
-
-        if model_class.__name__ in [
-            "FlaubertWithLMHeadModel",
-            "FunnelForPreTraining",
-            "ElectraForPreTraining",
-            "XLMWithLMHeadModel",
-        ]:
-            for k in key_differences:
-                if k in ["loss", "losses"]:
-                    tf_keys.discard(k)
-                    pt_keys.discard(k)
-        elif model_class.__name__.startswith("GPT2"):
-            # `TFGPT2` has `past_key_values` as a tensor while `GPT2` has it as a tuple.
-            tf_keys.discard("past_key_values")
-            pt_keys.discard("past_key_values")
-
-        # create new outputs from the remaining fields
-        new_tf_outputs = type(tf_outputs)(**{k: tf_outputs[k] for k in tf_keys})
-        new_pt_outputs = type(pt_outputs)(**{k: pt_outputs[k] for k in pt_keys})
-
-        return new_tf_outputs, new_pt_outputs
-
     def test_inputs_embeds(self):
         config, inputs_dict = self.model_tester.prepare_config_and_inputs_for_common()
 
@@ -3471,7 +3379,7 @@ class ModelTesterMixin:
                     )
 
     def test_model_is_small(self):
-        # Just a consistency check to make sure we are not running tests on 80M parameter models.
+        # Just a consistency check to make sure we are not running tests on 1M parameter models.
         config, _ = self.model_tester.prepare_config_and_inputs_for_common()
 
         for model_class in self.all_model_classes:
@@ -3975,57 +3883,6 @@ class ModelTesterMixin:
                 with torch.no_grad():
                     _ = model(**inputs_dict)
 
-    def test_sdpa_matches_eager_sliding_window(self):
-        if not self.has_attentions:
-            self.skipTest(reason="Model architecture does not support attentions")
-
-        WINDOW_ATTENTION_MODELS = ["mistral", "mixtral", "minimax", "qwen2", "qwen_moe", "starcoder2"]
-
-        if len(self.all_generative_model_classes) == 0:
-            self.skipTest(f"No generative model classes for {self.__class__.__name__}")
-
-        for model_class in self.all_generative_model_classes:
-            if model_class._supports_sdpa:
-                self.skipTest(reason="Model architecture does not support attentions")
-            config, inputs_dict = self.model_tester.prepare_config_and_inputs_for_common()
-
-            if config.model_type not in WINDOW_ATTENTION_MODELS:
-                self.skipTest(f"{config.model_type} does not use window attention")
-
-            config.sliding_window = 2
-
-            dummy_input = inputs_dict[model_class.main_input_name]
-            attention_mask = inputs_dict["attention_mask"]
-
-            self.assertTrue(dummy_input.ndim == 2)
-            self.assertTrue(dummy_input.shape[1] > 6)
-
-            with tempfile.TemporaryDirectory() as tmpdir:
-                with torch.device(torch_device):
-                    model_eager = AutoModelForCausalLM.from_config(
-                        config, attn_implementation="eager", dtype=torch.float32
-                    )
-
-                model_eager.save_pretrained(tmpdir)
-
-                with torch.device(torch_device):
-                    model_sdpa = AutoModelForCausalLM.from_pretrained(
-                        tmpdir, attn_implementation="sdpa", dtype=torch.float32
-                    )
-
-                model_eager = model_eager.eval()
-                model_sdpa = model_sdpa.eval()
-
-                with torch.no_grad():
-                    with sdpa_kernel(enable_flash=False, enable_math=True, enable_mem_efficient=False):
-                        res_eager = model_eager(**inputs_dict, return_dict=False)[0]
-                        res_sdpa = model_sdpa(**inputs_dict, return_dict=False)[0]
-
-                # Only non-padding tokens are expected to match.
-                self.assertTrue(
-                    torch.allclose(res_eager[attention_mask == 1], res_sdpa[attention_mask == 1], rtol=1e-4, atol=1e-4)
-                )
-
     def flash_attn_can_dispatch_composite_models(self, attn_implementation: str):
         """
         Tests if composite models can dispatch on flash attention if the sub-models support it.
@@ -4109,7 +3966,7 @@ class ModelTesterMixin:
         if not self.has_attentions:
             self.skipTest(reason="Model architecture does not support attentions")
 
-        for model_class in self.all_generative_model_classes:
+        for model_class in self.all_generative_model_classes:  # TODO: this test should run on all classes instead
             if not model_class._supports_flash_attn:
                 self.skipTest(f"{model_class.__name__} does not support Flash Attention 2")
             config, inputs_dict = self.model_tester.prepare_config_and_inputs_for_common()
@@ -4195,240 +4052,6 @@ class ModelTesterMixin:
 
         assert not loss.isnan().any()
 
-    def attention_mask_padding_matches_padding_free_with_position_ids(
-        self, attn_implementation: str, fa_kwargs: bool = False
-    ):
-        """
-        Tests that the given attention implementation can work with packed sequences and infers the mask
-        from position ids. This test requires the model to use new attention mask API which handles packing.
-        """
-        if not self.has_attentions:
-            self.skipTest(reason="Model architecture does not support attentions")
-
-        max_new_tokens = 30
-        support_flag = {
-            "sdpa": "_supports_sdpa",
-            "flash_attention_2": "_supports_flash_attn",
-            "flash_attention_3": "_supports_flash_attn",
-        }
-
-        for model_class in self.all_generative_model_classes:
-            if attn_implementation != "eager" and not getattr(model_class, support_flag[attn_implementation]):
-                self.skipTest(f"{model_class.__name__} does not support {attn_implementation}")
-
-            # can't infer if new attn mask API is supported by assume that only model with attention backend support it
-            if not model_class._supports_attention_backend:
-                self.skipTest(f"{model_class.__name__} does not support new attention mask API")
-
-            if model_class._is_stateful:  # non-transformer models most probably have no packing support
-                self.skipTest(f"{model_class.__name__} doesn't support packing!")
-
-            config, inputs_dict = self.model_tester.prepare_config_and_inputs_for_common()
-            if config.is_encoder_decoder:
-                self.skipTest("Model is an encoder-decoder")
-
-            if 0 not in inputs_dict.get("attention_mask", []) or "attention_mask" not in inputs_dict:
-                self.skipTest("Model dummy inputs should contain padding in their attention mask")
-
-            if "input_ids" not in inputs_dict or inputs_dict["input_ids"].ndim != 2:
-                self.skipTest("Model dummy inputs should contain text input ids")
-
-            # make sure that all models have enough positions for generation
-            dummy_input_ids = inputs_dict["input_ids"]
-            if hasattr(config, "max_position_embeddings"):
-                config.max_position_embeddings = max_new_tokens + dummy_input_ids.shape[1] + 1
-
-            model = model_class(config)
-            if "position_ids" not in inspect.signature(model.forward).parameters:
-                self.skipTest("Model does not support position_ids")
-
-            if (not fa_kwargs) and "position_ids" not in inspect.signature(model.forward).parameters:
-                continue  # this model doesn't accept position ids as input
-
-            with tempfile.TemporaryDirectory() as tmpdirname:
-                model.save_pretrained(tmpdirname)
-
-                # Drop all keys except for the minimal set. Hard to manipulate with multimodals/head_mask/etc
-                inputs_dict = {k: v for k, v in inputs_dict.items() if k in ["input_ids", "attention_mask"]}
-
-                # Ensure left padding, to adapt for some models
-                if 0 in inputs_dict["attention_mask"][:, -1]:
-                    inputs_dict["attention_mask"] = inputs_dict["attention_mask"].flip(1)
-                dummy_attention_mask = inputs_dict["attention_mask"]
-                dummy_input_ids[~dummy_attention_mask.bool()] = config.get_text_config().pad_token_id
-
-                model = (
-                    model_class.from_pretrained(
-                        tmpdirname,
-                        dtype=torch.bfloat16,
-                        attn_implementation=attn_implementation,
-                    )
-                    .to(torch_device)
-                    .eval()
-                )
-
-                if fa_kwargs:
-                    # flatten
-                    features = [
-                        {"input_ids": i[a.bool()].tolist()} for i, a in zip(dummy_input_ids, dummy_attention_mask)
-                    ]
-
-                    # add position_ids + fa_kwargs
-                    data_collator = DataCollatorWithFlattening(return_tensors="pt", return_flash_attn_kwargs=True)
-                    batch = data_collator(features)
-                    padfree_inputs_dict = {
-                        k: t.to(torch_device) if torch.is_tensor(t) else t for k, t in batch.items()
-                    }
-                else:
-                    # create packed position_ids
-                    position_ids = (
-                        torch.cat([torch.arange(length) for length in dummy_attention_mask.sum(1).tolist()])
-                        .long()
-                        .unsqueeze(0)
-                        .to(torch_device)
-                    )
-                    padfree_inputs_dict = {
-                        "input_ids": dummy_input_ids[dummy_attention_mask.bool()].unsqueeze(0),
-                        "position_ids": position_ids,
-                    }
-
-                # We need to do simple forward without cache in order to trigger packed SDPA/flex/eager attention path
-                res_padded = model(**inputs_dict, use_cache=False)
-                res_padfree = model(**padfree_inputs_dict, use_cache=False)
-
-                logits_padded = res_padded.logits[dummy_attention_mask.bool()]
-                logits_padfree = res_padfree.logits[0]
-
-                # acceptable numerical instability
-                tol = torch.finfo(torch.bfloat16).eps
-                torch.testing.assert_close(logits_padded, logits_padfree, rtol=tol, atol=tol)
-
-    def test_eager_padding_matches_padding_free_with_position_ids(self):
-        self.attention_mask_padding_matches_padding_free_with_position_ids(attn_implementation="eager")
-
-    def test_sdpa_padding_matches_padding_free_with_position_ids(self):
-        self.attention_mask_padding_matches_padding_free_with_position_ids(attn_implementation="sdpa")
-
-    @require_flash_attn
-    @require_torch_gpu
-    @mark.flash_attn_test
-    @slow
-    def test_flash_attention_2_padding_matches_padding_free_with_position_ids(self):
-        self.attention_mask_padding_matches_padding_free_with_position_ids(attn_implementation="flash_attention_2")
-
-    @require_flash_attn
-    @require_torch_gpu
-    @mark.flash_attn_test
-    @slow
-    def test_flash_attention_2_padding_matches_padding_free_with_position_ids_and_fa_kwargs(self):
-        self.attention_mask_padding_matches_padding_free_with_position_ids(
-            attn_implementation="flash_attention_2", fa_kwargs=True
-        )
-
-    @require_flash_attn_3
-    @require_torch_gpu
-    @mark.flash_attn_3_test
-    @slow
-    def test_flash_attention_3_padding_matches_padding_free_with_position_ids(self):
-        self.attention_mask_padding_matches_padding_free_with_position_ids(attn_implementation="flash_attention_3")
-
-    @require_flash_attn_3
-    @require_torch_gpu
-    @mark.flash_attn_3_test
-    @slow
-    def test_flash_attention_3_padding_matches_padding_free_with_position_ids_and_fa_kwargs(self):
-        self.attention_mask_padding_matches_padding_free_with_position_ids(
-            attn_implementation="flash_attention_3", fa_kwargs=True
-        )
-
-    @require_flash_attn
-    @require_torch_gpu
-    @mark.flash_attn_test
-    def test_flash_attention_2_continue_generate_with_position_ids(self):
-        """
-        Tests whether flash attention can continue its generation from given position ids.
-
-        NOTE: This serves as regression check as we had instances where flash attention entered the varlen
-        path here. It should now always enter the base `flash_fn`.
-        """
-
-        max_new_tokens = 2
-        for model_class in self.all_generative_model_classes:
-            if not model_class._supports_flash_attn:
-                self.skipTest(f"{model_class.__name__} does not support Flash Attention.")
-
-            config, inputs_dict = self.model_tester.prepare_config_and_inputs_for_common()
-            if config.is_encoder_decoder:
-                self.skipTest("Model is an encoder-decoder")
-
-            if not hasattr(config.get_text_config(), "use_cache"):
-                self.skipTest(f"{model_class.__name__} doesn't support caching")
-
-            if "input_ids" not in inputs_dict or inputs_dict["input_ids"].ndim != 2:
-                self.skipTest("Model dummy inputs should contain text input ids")
-
-            # make sure that all models have enough positions for generation
-            dummy_input_ids = inputs_dict["input_ids"]
-            if hasattr(config, "max_position_embeddings"):
-                config.max_position_embeddings = max_new_tokens + dummy_input_ids.shape[1] + 1
-
-            model = model_class(config)
-            if "position_ids" not in inspect.signature(model.forward).parameters:
-                self.skipTest("Model does not support position_ids")
-
-            with tempfile.TemporaryDirectory() as tmpdirname:
-                model.save_pretrained(tmpdirname)
-                model = (
-                    model_class.from_pretrained(
-                        tmpdirname,
-                        dtype=torch.bfloat16,
-                        attn_implementation="flash_attention_2",
-                    )
-                    .to(torch_device)
-                    .eval()
-                )
-
-                # Drop all keys except for `input_ids`. Hard to manipulate with multimodals/head_mask/etc
-                dummy_input_ids = inputs_dict["input_ids"]
-                dummy_position_ids = torch.arange(dummy_input_ids.shape[1], device=torch_device)
-                dummy_position_ids = dummy_position_ids.unsqueeze(0).repeat(dummy_input_ids.shape[0], 1)
-
-                # Store cache for the input prompt
-                output = model(dummy_input_ids, position_ids=dummy_position_ids, use_cache=True)
-                if "past_key_values" not in output:
-                    self.skipTest("This model doesn't return `past_key_values`")
-
-                # create new input_ids and position_ids to continue generation re-using the cache
-                new_input_ids = output.logits[:, -1, :].float().argmax(-1)[:, None]
-                past_length = dummy_input_ids.shape[1]
-                position_ids = torch.arange(past_length, past_length + new_input_ids.shape[1], device=torch_device)
-                position_ids = position_ids.unsqueeze(0).repeat(new_input_ids.shape[0], 1)
-
-                output = model(
-                    input_ids=new_input_ids,
-                    past_key_values=output.past_key_values,
-                    position_ids=position_ids,
-                    use_cache=True,
-                )
-                next_token_logits = output.logits[:, -1, :].float()
-
-                generate_kwargs = {
-                    "pad_token_id": -1,
-                    "eos_token_id": -1,
-                    "forced_eos_token_id": None,
-                    "use_cache": True,
-                    "do_sample": False,
-                    "return_dict_in_generate": True,
-                    "output_logits": True,
-                    "max_new_tokens": max_new_tokens,
-                }
-                generation_out = model.generate(dummy_input_ids, **generate_kwargs)
-                next_token_logits_from_generate = generation_out.logits[-1]
-
-                # acceptable numerical instability
-                tol = torch.finfo(torch.bfloat16).eps
-                torch.testing.assert_close(next_token_logits_from_generate, next_token_logits, rtol=tol, atol=tol)
-
     def flash_attn_from_config(self, attn_implementation: str):
         r"""
         Tests if the model can be loaded with `attn_implementation` from the config and if the
@@ -4437,7 +4060,7 @@ class ModelTesterMixin:
         if not self.has_attentions:
             self.skipTest(reason="Model architecture does not support attentions")
 
-        for model_class in self.all_generative_model_classes:
+        for model_class in self.all_generative_model_classes:  # TODO: this test should run on all classes instead
             if not model_class._supports_flash_attn:
                 self.skipTest(f"{model_class.__name__} does not support {attn_implementation}")
 
@@ -4482,41 +4105,6 @@ class ModelTesterMixin:
     @slow
     def test_flash_attn_3_from_config(self):
         self.flash_attn_from_config(attn_implementation="flash_attention_3")
-
-    def _get_custom_4d_mask_test_data(self):
-        # Sequence in which all but the last token is the same
-        input_ids = torch.tensor(
-            [[10, 11, 12, 13], [10, 11, 12, 14], [10, 11, 12, 15]], device=torch_device, dtype=torch.int64
-        )
-        position_ids = torch.tensor([[0, 1, 2, 3]] * 3, device=torch_device, dtype=torch.int64)
-
-        # Combining common prefix with the unique ending tokens:
-        input_ids_shared_prefix = torch.cat([input_ids[0][:-1], input_ids[:, -1]]).unsqueeze(0)
-
-        # Creating a 4D mask where each of the last 3 tokens do not attend to each other.
-        mask_shared_prefix = torch.tensor(
-            [
-                [
-                    [
-                        [1, 0, 0, 0, 0, 0],
-                        [1, 1, 0, 0, 0, 0],
-                        [1, 1, 1, 0, 0, 0],
-                        [1, 1, 1, 1, 0, 0],
-                        [1, 1, 1, 0, 1, 0],
-                        [1, 1, 1, 0, 0, 1],
-                    ]
-                ]
-            ],
-        )
-        # inverting the attention mask
-        mask_dtype = torch.float32
-        min_dtype = torch.finfo(mask_dtype).min
-        mask_shared_prefix = (mask_shared_prefix.eq(0.0)).to(dtype=mask_dtype, device=torch_device) * min_dtype
-
-        # Creating a position_ids tensor. note the repeating figures in the end.
-        position_ids_shared_prefix = torch.tensor([[0, 1, 2, 3, 3, 3]], device=torch_device, dtype=torch.int64)
-
-        return input_ids, position_ids, input_ids_shared_prefix, mask_shared_prefix, position_ids_shared_prefix
 
     def test_sliding_window_mask(self):
         """Tests that we can control the sliding window attention behavior of a model."""
@@ -4572,58 +4160,6 @@ class ModelTesterMixin:
             for layer_attention in attentions_not_sliding:
                 self.assertFalse((layer_attention[:, :, ~sliding_mask] == 0).all().item())
 
-    def test_custom_4d_attention_mask(self):
-        if not self.has_attentions:
-            self.skipTest(reason="Model architecture does not support attentions")
-
-        if len(self.all_generative_model_classes) == 0:
-            self.skipTest(
-                reason="Model architecture has no generative classes, and thus not necessarily supporting 4D masks"
-            )
-
-        set_model_tester_for_less_flaky_test(self)
-
-        for model_class in self.all_generative_model_classes:
-            if not model_class._can_compile_fullgraph:
-                self.skipTest(f"{model_class.__name__} is not guaranteed to work with custom 4D attention masks")
-            config, _ = self.model_tester.prepare_config_and_inputs_for_common()
-            set_config_for_less_flaky_test(config)
-            if getattr(config, "sliding_window", 0) is not None and getattr(config, "sliding_window", 0) > 0:
-                self.skipTest(f"{model_class.__name__} with sliding window attention is not supported by this test")
-            model = model_class(config).to(device=torch_device, dtype=torch.float32).eval()
-            set_model_for_less_flaky_test(model)
-            if "position_ids" not in inspect.signature(model.forward).parameters:
-                continue  # model doesn't accept position ids and probably has special way to model positions
-
-            if "position_ids" not in inspect.signature(model.forward).parameters:
-                continue  # this model doesn't accept position ids as input
-
-            (
-                input_ids,
-                position_ids,
-                input_ids_shared_prefix,
-                mask_shared_prefix,
-                position_ids_shared_prefix,
-            ) = self._get_custom_4d_mask_test_data()
-
-            logits = model.forward(input_ids, position_ids=position_ids).logits
-            # logits.shape == torch.Size([3, 4, ...])
-
-            logits_shared_prefix = model(
-                input_ids_shared_prefix,
-                attention_mask=mask_shared_prefix,
-                position_ids=position_ids_shared_prefix,
-            )[0]
-            # logits_shared_prefix.shape == torch.Size([1, 6, ...])
-
-            out_last_tokens = logits[:, -1, :]  # last tokens in each batch line
-            out_shared_prefix_last_tokens = logits_shared_prefix[0, -3:, :]  # last three tokens
-
-            # comparing softmax-normalized logits:
-            normalized_0 = F.softmax(out_last_tokens, dim=-1)
-            normalized_1 = F.softmax(out_shared_prefix_last_tokens, dim=-1)
-            torch.testing.assert_close(normalized_0, normalized_1, rtol=1e-3, atol=1e-3)
-
     @slow
     @require_torch_accelerator
     @pytest.mark.torch_compile_test
@@ -4673,29 +4209,6 @@ class ModelTesterMixin:
         # check grad matches
         for name, param in model._orig_mod.named_parameters():
             torch.testing.assert_close(param.grad.detach().cpu(), params[name], rtol=1e-4, atol=1e-4)
-
-    def test_forward_with_logits_to_keep(self):
-        for model_class in self.all_generative_model_classes:
-            if "logits_to_keep" not in set(inspect.signature(model_class.forward).parameters.keys()):
-                self.skipTest(reason="This model does not support `logits_to_keep` argument.")
-
-            config, inputs = self.model_tester.prepare_config_and_inputs_for_common()
-            batch_size, sequence_length = inputs["input_ids"].shape[:2]
-            vocab_size = config.get_text_config().vocab_size
-            model = model_class(config).to(device=torch_device).eval()
-            # some models have labels but `logits_to_keep` should not be used in train mode
-            _ = inputs.pop("labels", None)
-
-            # logits_to_keep=0 is a special case meaning "keep all logits"
-            all_logits = model(**inputs, logits_to_keep=0).logits
-            last_token_logits = model(**inputs, logits_to_keep=1).logits
-
-            # Assert all shapes are correct
-            self.assertEqual(tuple(all_logits.shape), (batch_size, sequence_length, vocab_size))
-            self.assertEqual(tuple(last_token_logits.shape), (batch_size, 1, vocab_size))
-
-            # Assert the last tokens are actually the same (except for the natural fluctuation due to order of FP ops)
-            torch.testing.assert_close(all_logits[:, -1:, :], last_token_logits, rtol=1e-5, atol=1e-5)
 
     @slow
     @require_torch_greater_or_equal("2.5")
